Refactor nanoGPT runtime modules
