+++ package-files.test.ts
@@ -32,6 +32,12 @@ describe("package manifest files", () => {
         "api.ts",
         "models.ts",
         "runtime.ts",
+        "runtime/config.ts",
+        "runtime/dynamic-models.ts",
+        "runtime/discovery.ts",
+        "runtime/provider-pricing.ts",
+        "runtime/routing.ts",
+        "runtime/usage.ts",
         "provider-catalog.ts",
         "provider-discovery.ts",
         "image-generation-provider.ts",
