+++ package.json
@@ -24,6 +24,12 @@
     "api.ts",
     "models.ts",
     "runtime.ts",
+    "runtime/config.ts",
+    "runtime/dynamic-models.ts",
+    "runtime/discovery.ts",
+    "runtime/provider-pricing.ts",
+    "runtime/routing.ts",
+    "runtime/usage.ts",
     "provider-catalog.ts",
     "provider-discovery.ts",
     "nanogpt-errors.ts",
